Implement GroundingSpace using AtomIndex
Implement everything except removing atom. Also AtomIndex doesn't keep
and return duplicated atom.
